refactor: implement plan, use manager-wide command runner
This is a large structural change to how concierge goes about its
business, but it removes a fair bit of code duplication.

When invoked either with `prepare` or `restore`, the tool will now
generate a `Plan`, which comprises:

- A list of snaps
- A list of debs
- A list of providers
- Juju configuration

Snap installation is now taken care of centrally - providers no longer
install/remove their own snaps, but instead report which snaps they
need at the time the `Plan` is built.

There are a set of `Handler`s which each contain the logic for
installing/removing/configuring a thing or set of things.

Each manager now has a `runner` property, which is passed to each
handler, and makes for easier central control of log output etc.