feat: add `bootstrap` option to providers
This commit implements a new `bootstrap` option for providers.

This is because in some cases, it's desirable to have a provider
installed, but not bootstrapped with a Juju controller.

The most clear example of this is when building a K8s charm, LXD
is still required to build the charm with charmcraft.