feat!: require `concierge` to be run as `root`
Before this commit, nearly all commands were run using `sudo`.
This was making the assumption that the user executing the
command was able to escalate using `sudo` with no password.
There was no handling in `concierge` to enable a password prompt,
so if the relevant configuration was not in place for NOPASSWD
sudo, then `concierge` would hang.

Additionally, by not being able to gurantee root privileges,
`concierge` was unable to take full advantage of the native
`snapd` client package for installing multiple snaps.