refactor: implement the `SnapPackage` interface
Because the `packages.Snap` methods rely on speaking to `snapd` in order
to determine install status, confinement of a snap and default channel,
it makes for difficult testing.

This commit introduces a new interface which can later be implemented
by a test fixture, which will enable testing the flow of the
`SnapHandler` without relying on `snapd` at test time.