refactor: centralise tracking of real user to the runner
This commit centralises the user tracking onto the runner.
Previously, there were multiple places where the real user,
home directory and other attributes were calculated.

Now, the effective user is an attribute of the runner, which
also simplifies testing with the mock runner, and removes a
lot of the dynamism from the unit tests.

Additionally, this commit wraps filesystem operations with the
runner, which has the advantage of making them easier to test.