feat: add `runner.RunExclusive`
`runner.RunExclusive` is used to execute commands that cannot be invoked
in parallel, for example when multiple providers are trying to install
snaps.

The implementation is based on a map of mutexes, where the key is the
executable name. When `RunExclusive` is invoked, the executable is used
to lookup the mutex (creating if it doesn't exist), then locks it
before running the command.

Once the command has returned the mutex in unlocked.