refactor: providers install their own snaps
This change shifts the responsibility for installing provider/juju snaps
from the manager/general `SnapHandler` onto the provider.

This means that as `Prepare()` is called on each provider, a new local
`SnapHandler` is established using the central runner to install the
snaps. This makes the first proper use of `runner.RunExclusive` since
the providers are initialised in parallel.

Lay the ground work for providers having credentials by modifying the
provider interface to include a `Credentials` method.