docs: clarify functionality of restore command
Previously, documentation explained the restore command as a
command that would restore a machine to it's pre-provisioned
state.

This was not quite accurate, in that 'restore' would perform
removal of snaps/debs/config in the plan irrespective of whether
they happened to be in place prior to running 'prepare'.

I could have changed this behaviour, but I think that would lead
to potential inconsitencies in how `concierge` behaves depending
on where it is run (and the machine state prior to running it),
which somewhat defeats the purpose - so instead I have clarified
the documentation.

Closes #22.