fix(k8s): specify alternate `containerd-base-dir`
On machines where containerd is already present
the k8s snap can fail to bootstrap due to a
conflict in the path.

Github Actions runners ship with containerd by
default, and they account for a large percentage
of the machines where concierge will be run. As a
result, this change will configure that base path
to an alternate path to avoid the clash.