feat: add the ability to disable juju during `concierge prepare`
This commit adds the ability through config, command line flag, or env
var, to disable the Juju handler.

There are some situations where this is desirable, such as in CI
workflows that *only* need to build artifacts, but don't need Juju.

Because `concierge` is primarily a tool for setting up charm testing
environments, unlike the providers configs, the Juju option is named
`juju.disable`, rather than `juju.enable`. This will prevent the
addition of this option breaking any previous configurations already in
use.