fix: retry bootstrap check on on controllers when there are errors
At the moment, when concierge checks for a bootstrapped controller, if
the command fails the check is considered a failure, and thus that the
controller doesn't exist.

In some cases, that is correct, and we can tell that from the error
message. In other cases, the check fails because the controller fails
to respond.

This commit introduces a retry with backoff on that check.